ADJST-671 Moving unused deductions into the adjustments api
