--- src/main/kotlin/uk/gov/justice/digital/hmpps/adjustments/api/client/SystemPrisonApiClient.kt
@@ -1,23 +1,0 @@
-package uk.gov.justice.digital.hmpps.adjustments.api.client
-
-import org.slf4j.LoggerFactory
-import org.springframework.beans.factory.annotation.Qualifier
-import org.springframework.core.ParameterizedTypeReference
-import org.springframework.stereotype.Service
-import org.springframework.web.reactive.function.client.WebClient
-import uk.gov.justice.digital.hmpps.adjustments.api.model.prisonapi.PrisonerDetails
-
-@Service
-class SystemPrisonApiClient(@Qualifier("systemPrisonApiWebClient") private val webClient: WebClient) {
-  private inline fun <reified T> typeReference() = object : ParameterizedTypeReference<T>() {}
-  private val log = LoggerFactory.getLogger(this::class.java)
-
-  fun getPrisonerDetail(nomsId: String): PrisonerDetails {
-    log.info("Requesting details for nomsId $nomsId")
-    return webClient.get()
-      .uri("/api/offenders/$nomsId")
-      .retrieve()
-      .bodyToMono(typeReference<PrisonerDetails>())
-      .block()!!
-  }
-}
