+++ src/test/kotlin/uk/gov/justice/digital/hmpps/adjustments/api/service/UnusedDeductionsEventServiceTest.kt
@@ -17,46 +17,31 @@ import uk.gov.justice.digital.hmpps.adjustments.api.model.RemandDto
 import uk.gov.justice.digital.hmpps.adjustments.api.model.TaggedBailDto
 import uk.gov.justice.digital.hmpps.adjustments.api.model.UnusedDeductionCalculationResponse
 import java.time.LocalDate
-import java.time.LocalDateTime
 import java.util.UUID
 
 class UnusedDeductionsEventServiceTest {
 
   private val adjustmentService = mock<AdjustmentsService>()
   private val calculateReleaseDatesApiClient = mock<CalculateReleaseDatesApiClient>()
-  private val adjustmentsDomainEventService = mock<AdjustmentsDomainEventService>()
 
   private val unusedDeductionsService = UnusedDeductionsService(
     adjustmentService,
     calculateReleaseDatesApiClient,
-    adjustmentsDomainEventService
   )
 
   private val eventService = UnusedDeductionsEventService(
     unusedDeductionsService,
   )
-  /*
-
-  @Schema(description = "The details of a remand adjustment")
-  val remand: RemandDto?,
-  @Schema(description = "The details of an additional days awarded adjustments (ADA)")
-  val additionalDaysAwarded: AdditionalDaysAwardedDto?,
-  @Schema(description = "Additional details of a UAL adjustment")
-  val unlawfullyAtLarge: UnlawfullyAtLargeDto?,
-  @Schema(description = "The details of a tagged-bail adjustment")
-  val taggedBail: TaggedBailDto?,
-  @Schema(description = "The NOMIS sentence sequence of the adjustment")
-  val sentenceSequence: Int? = null,
-   */
 
   val person = "ABC123"
   private val ninetyDaysRemand = AdjustmentDto(
     UUID.randomUUID(), 1, person, AdjustmentType.REMAND, LocalDate.now().minusDays(100),
-    LocalDate.now().minusDays(9),  90,  remand = RemandDto(listOf(1L)), additionalDaysAwarded = null, unlawfullyAtLarge = null, taggedBail = null,
+    LocalDate.now().minusDays(9), 90, remand = RemandDto(listOf(1L)), additionalDaysAwarded = null, unlawfullyAtLarge = null, taggedBail = null,
     effectiveDays = 90,
     sentenceSequence = 1,
     source = AdjustmentSource.DPS,
   )
+
   @Nested
   inner class AdjustmentEventTests {
     @Test
@@ -76,7 +61,7 @@ class UnusedDeductionsEventServiceTest {
       )
       val adjustments = listOf(remand, taggedBail, unusedDeductions)
 
-      whenever(adjustmentService.findCurrentAdjustments(person, AdjustmentStatus.ACTIVE)).thenReturn(adjustments)
+      whenever(adjustmentService.findCurrentAdjustments(person, AdjustmentStatus.ACTIVE, null)).thenReturn(adjustments)
       whenever(calculateReleaseDatesApiClient.calculateUnusedDeductions(adjustments, person)).thenReturn(
         UnusedDeductionCalculationResponse(100),
       )
@@ -109,7 +94,7 @@ class UnusedDeductionsEventServiceTest {
       )
       val adjustments = listOf(remand, taggedBail)
 
-      whenever(adjustmentsApiClient.getAdjustmentsByPerson(person)).thenReturn(adjustments)
+      whenever(adjustmentService.findCurrentAdjustments(person, AdjustmentStatus.ACTIVE, null)).thenReturn(adjustments)
       whenever(calculateReleaseDatesApiClient.calculateUnusedDeductions(adjustments, person)).thenReturn(
         UnusedDeductionCalculationResponse(100),
       )
@@ -125,15 +110,17 @@ class UnusedDeductionsEventServiceTest {
         ),
       )
 
-      verify(adjustmentsApiClient).updateEffectiveDays(AdjustmentEffectiveDays(taggedBail.id!!, 80, person))
-      verify(adjustmentsApiClient).updateEffectiveDays(AdjustmentEffectiveDays(remand.id!!, 0, person))
-      verify(adjustmentsApiClient).createAdjustment(
-        remand.copy(
-          id = null,
-          toDate = null,
-          fromDate = null,
-          days = 100,
-          adjustmentType = AdjustmentType.UNUSED_DEDUCTIONS,
+      verify(adjustmentService).updateEffectiveDays(taggedBail.id!!, AdjustmentEffectiveDaysDto(taggedBail.id!!, 80, person))
+      verify(adjustmentService).updateEffectiveDays(remand.id!!, AdjustmentEffectiveDaysDto(remand.id!!, 0, person))
+      verify(adjustmentService).create(
+        listOf(
+          remand.copy(
+            id = null,
+            toDate = null,
+            fromDate = null,
+            days = 100,
+            adjustmentType = AdjustmentType.UNUSED_DEDUCTIONS,
+          ),
         ),
       )
     }
@@ -142,7 +129,7 @@ class UnusedDeductionsEventServiceTest {
     fun updateUnusedDeductions_ZeroCalculatedDays() {
       val person = "ABC123"
       val remand = ninetyDaysRemand.copy(
-        effectiveDays = 80
+        effectiveDays = 80,
       )
       val taggedBail = remand.copy(
         id = UUID.randomUUID(),
@@ -158,7 +145,7 @@ class UnusedDeductionsEventServiceTest {
       )
       val adjustments = listOf(remand, taggedBail, unusedDeductions)
 
-      whenever(adjustmentsApiClient.getAdjustmentsByPerson(person)).thenReturn(adjustments)
+      whenever(adjustmentService.findCurrentAdjustments(person, AdjustmentStatus.ACTIVE, null)).thenReturn(adjustments)
       whenever(calculateReleaseDatesApiClient.calculateUnusedDeductions(adjustments, person)).thenReturn(
         UnusedDeductionCalculationResponse(0),
       )
@@ -174,20 +161,20 @@ class UnusedDeductionsEventServiceTest {
         ),
       )
 
-      verify(adjustmentsApiClient).updateEffectiveDays(AdjustmentEffectiveDays(taggedBail.id!!, 90, person))
-      verify(adjustmentsApiClient).updateEffectiveDays(AdjustmentEffectiveDays(remand.id!!, 90, person))
-      verify(adjustmentsApiClient).deleteAdjustment(unusedDeductions.id!!)
+      verify(adjustmentService).updateEffectiveDays(taggedBail.id!!, AdjustmentEffectiveDaysDto(taggedBail.id!!, 90, person))
+      verify(adjustmentService).updateEffectiveDays(remand.id!!, AdjustmentEffectiveDaysDto(remand.id!!, 90, person))
+      verify(adjustmentService).delete(unusedDeductions.id!!)
     }
 
     @Test
     fun updateUnusedDeductions_NoDeductionsButUnusedDeductions() {
       val person = "ABC123"
       val unusedDeductions = ninetyDaysRemand.copy(
-        adjustmentType = AdjustmentType.UNUSED_DEDUCTIONS
+        adjustmentType = AdjustmentType.UNUSED_DEDUCTIONS,
       )
       val adjustments = listOf(unusedDeductions)
 
-      whenever(adjustmentsApiClient.getAdjustmentsByPerson(person)).thenReturn(adjustments)
+      whenever(adjustmentService.findCurrentAdjustments(person, AdjustmentStatus.ACTIVE, null)).thenReturn(adjustments)
       whenever(calculateReleaseDatesApiClient.calculateUnusedDeductions(adjustments, person)).thenReturn(
         UnusedDeductionCalculationResponse(0),
       )
@@ -203,7 +190,7 @@ class UnusedDeductionsEventServiceTest {
         ),
       )
 
-      verify(adjustmentsApiClient).deleteAdjustment(unusedDeductions.id!!)
+      verify(adjustmentService).delete(unusedDeductions.id!!)
     }
 
     @Test
@@ -220,7 +207,7 @@ class UnusedDeductionsEventServiceTest {
         ),
       )
 
-      verifyNoInteractions(adjustmentsApiClient)
+      verifyNoInteractions(adjustmentService)
     }
   }
 
@@ -228,6 +215,44 @@ class UnusedDeductionsEventServiceTest {
   inner class GeneralTests {
     @Test
     fun `Do not update unused deductions if adjustments arent DPS`() {
+      val person = "ABC123"
+      val remand = ninetyDaysRemand.copy(source = AdjustmentSource.NOMIS)
+      val taggedBail = remand.copy(
+        id = UUID.randomUUID(),
+        adjustmentType = AdjustmentType.TAGGED_BAIL,
+        taggedBail = TaggedBailDto(1),
+        remand = null,
+      )
+      val unusedDeductions = remand.copy(
+        id = UUID.randomUUID(),
+        adjustmentType = AdjustmentType.UNUSED_DEDUCTIONS,
+        days = 10,
+        effectiveDays = 10,
+      )
+      val adjustments = listOf(remand, taggedBail, unusedDeductions)
+
+      whenever(adjustmentService.findCurrentAdjustments(person, AdjustmentStatus.ACTIVE, null)).thenReturn(adjustments)
+
+      eventService.handleAdjustmentMessage(
+        AdjustmentEvent(
+          AdjustmentAdditionalInformation(
+            id = UUID.randomUUID().toString(),
+            offenderNo = person,
+            source = "DPS",
+            false,
+          ),
+        ),
+      )
+
+      verify(adjustmentService).findCurrentAdjustments(person, AdjustmentStatus.ACTIVE, null)
+      verifyNoMoreInteractions(adjustmentService)
+    }
+  }
+
+  @Nested
+  inner class PrisonerSearchTests {
+    @Test
+    fun `Handle prisoner event`() {
       val person = "ABC123"
       val remand = ninetyDaysRemand.copy()
       val taggedBail = remand.copy(
@@ -244,49 +269,7 @@ class UnusedDeductionsEventServiceTest {
       )
       val adjustments = listOf(remand, taggedBail, unusedDeductions)
 
-      whenever(adjustmentsApiClient.getAdjustmentsByPerson(person)).thenReturn(adjustments)
-
-      eventService.handleAdjustmentMessage(
-        AdjustmentEvent(
-          AdjustmentAdditionalInformation(
-            id = UUID.randomUUID().toString(),
-            offenderNo = person,
-            source = "DPS",
-            false,
-          ),
-        ),
-      )
-
-      verify(adjustmentsApiClient).getAdjustmentsByPerson(person)
-      verifyNoMoreInteractions(adjustmentsApiClient)
-    }
-  }
-
-  @Nested
-  inner class PrisonerSearchTests {
-    @Test
-    fun `Handle prisoner event`() {
-      val person = "ABC123"
-      val remand = Adjustment(
-        UUID.randomUUID(), 1, 1, person, AdjustmentType.REMAND, LocalDate.now().minusDays(100),
-        LocalDate.now().minusDays(9), LocalDateTime.now(), 90, 90, taggedBail = null, remand = RemandDto(listOf(1L)),
-        source = AdjustmentSource.DPS,
-      )
-      val taggedBail = remand.copy(
-        id = UUID.randomUUID(),
-        adjustmentType = AdjustmentType.TAGGED_BAIL,
-        taggedBail = TaggedBailDto(1),
-        remand = null,
-      )
-      val unusedDeductions = remand.copy(
-        id = UUID.randomUUID(),
-        adjustmentType = AdjustmentType.UNUSED_DEDUCTIONS,
-        days = 10,
-        effectiveDays = 10,
-      )
-      val adjustments = listOf(remand, taggedBail, unusedDeductions)
-
-      whenever(adjustmentsApiClient.getAdjustmentsByPerson(person)).thenReturn(adjustments)
+      whenever(adjustmentService.findCurrentAdjustments(person, AdjustmentStatus.ACTIVE, null)).thenReturn(adjustments)
       whenever(calculateReleaseDatesApiClient.calculateUnusedDeductions(adjustments, person)).thenReturn(
         UnusedDeductionCalculationResponse(100),
       )
@@ -300,9 +283,9 @@ class UnusedDeductionsEventServiceTest {
         ),
       )
 
-      verify(adjustmentsApiClient).updateEffectiveDays(AdjustmentEffectiveDays(taggedBail.id!!, 80, person))
-      verify(adjustmentsApiClient).updateEffectiveDays(AdjustmentEffectiveDays(remand.id!!, 0, person))
-      verify(adjustmentsApiClient).updateAdjustment(unusedDeductions.copy(days = 100))
+      verify(adjustmentService).updateEffectiveDays(taggedBail.id!!, AdjustmentEffectiveDaysDto(taggedBail.id!!, 80, person))
+      verify(adjustmentService).updateEffectiveDays(remand.id!!, AdjustmentEffectiveDaysDto(remand.id!!, 0, person))
+      verify(adjustmentService).update(unusedDeductions.id!!, unusedDeductions.copy(days = 100))
     }
 
     @Test
@@ -318,7 +301,7 @@ class UnusedDeductionsEventServiceTest {
         ),
       )
 
-      verifyNoInteractions(adjustmentsApiClient)
+      verifyNoInteractions(adjustmentService)
     }
   }
 }
